// src/services/central-api/generated/model/crt-vehicle-master-vehicle-master-search-output-pagination.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PL Cental API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { VehicleMasterSearchOutput } from './vehicle-master-search-output';

/**
 * 
 * @export
 * @interface CrtVehicleMasterVehicleMasterSearchOutputPagination
 */
export interface CrtVehicleMasterVehicleMasterSearchOutputPagination {
    /**
     * 
     * @type {Array<VehicleMasterSearchOutput>}
     * @memberof CrtVehicleMasterVehicleMasterSearchOutputPagination
     */
    'items'?: Array<VehicleMasterSearchOutput> | null;
    /**
     * 
     * @type {number}
     * @memberof CrtVehicleMasterVehicleMasterSearchOutputPagination
     */
    'totalItems'?: number;
    /**
     * 
     * @type {number}
     * @memberof CrtVehicleMasterVehicleMasterSearchOutputPagination
     */
    'pageIndex'?: number;
    /**
     * 
     * @type {number}
     * @memberof CrtVehicleMasterVehicleMasterSearchOutputPagination
     */
    'pageSize'?: number;
    /**
     * 
     * @type {number}
     * @memberof CrtVehicleMasterVehicleMasterSearchOutputPagination
     */
    'pageTotal'?: number;
}